Replace V1 IDA mapping generator with V2 (renamed to IDAMappingGenerator, outputs to IDAMappings)
