+++ Dumper/Generator/Public/SharedPredefinedMembers.h
@@ -3,7 +3,7 @@
 #include "PredefinedMembers.h"
 
 /*
-* Initializes predefined members for core engine types shared between CppGenerator and IDAMappingV2Generator
+* Initializes predefined members for core engine types shared between CppGenerator and IDAMappingGenerator
 * Registers members for UObject, UField, UEnum, UStruct, UFunction, UClass, ULevel, UDataTable
 */
 void InitCorePredefinedMembers(PredefinedMemberLookupMapType& OutMembers);
