+++ Dumper/main.cpp
@@ -7,7 +7,6 @@
 #include "Generators/MappingGenerator.h"
 #include "Generators/IDAMappingGenerator.h"
 #include "Generators/DumpspaceGenerator.h"
-#include "Generators/IDAMappingV2Generator.h"
 
 #include "Generators/Generator.h"
 
@@ -63,7 +62,6 @@ DWORD MainThread(HMODULE Module)
 	Generator::Generate<MappingGenerator>();
 	Generator::Generate<IDAMappingGenerator>();
 	Generator::Generate<DumpspaceGenerator>();
-	Generator::Generate<IDAMappingV2Generator>();
 
 	auto DumpFinishTime = std::chrono::high_resolution_clock::now();
